refactor #56 PasswordEncryptor 를 빈으로 등록해서 필요시 도메인 객체에 주입한다.
