+++ api/member-api-list.http
@@ -1,0 +1,7 @@
+POST localhost:8080/login
+Content-Type: application/json
+
+{
+  "email": "mhoh@miridi.com",
+  "password" : "test1234"
+}
