+++ src/main/java/com/squadb/workassistantapi/service/PasswordEncryptorImpl.java
@@ -1,7 +1,8 @@
-package com.squadb.workassistantapi.util;
+package com.squadb.workassistantapi.service;
 
-import lombok.experimental.UtilityClass;
+import com.squadb.workassistantapi.domain.PasswordEncryptor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Component;
 
 import javax.crypto.SecretKeyFactory;
 import javax.crypto.spec.PBEKeySpec;
@@ -12,8 +13,8 @@ import java.security.spec.InvalidKeySpecException;
 import java.security.spec.KeySpec;
 
 @Slf4j
-@UtilityClass
-public class HashUtil {
+@Component
+public class PasswordEncryptorImpl implements PasswordEncryptor {
 
     private static final String SECRET_KEY_ALGORITHM = "PBKDF2WithHmacSHA1";
     private static final int ITERATION_COUNT = 1000;
@@ -21,14 +22,15 @@ public class HashUtil {
     private static final int SALT_BYTES = 16;
     private static final String HASHED_PASSWORD_DELIMITER = ":";
 
-    public static String hashPassword(String plainPassword) {
+    @Override
+    public String encrypt(String plain) {
         final SecureRandom random = new SecureRandom();
         final byte[] salt = new byte[SALT_BYTES];
         random.nextBytes(salt);
 
         try {
             final SecretKeyFactory factory = SecretKeyFactory.getInstance(SECRET_KEY_ALGORITHM);
-            final KeySpec spec = new PBEKeySpec(plainPassword.toCharArray(), salt, ITERATION_COUNT, KEY_LENGTH);
+            final KeySpec spec = new PBEKeySpec(plain.toCharArray(), salt, ITERATION_COUNT, KEY_LENGTH);
             final byte[] hash = factory.generateSecret(spec).getEncoded();
 
             return String.join(HASHED_PASSWORD_DELIMITER, SECRET_KEY_ALGORITHM, toHex(salt), String.valueOf(ITERATION_COUNT), String.valueOf(KEY_LENGTH), toHex(hash));
@@ -38,15 +40,16 @@ public class HashUtil {
         }
     }
 
-    public static boolean equalPassword(String passwordInput, String hashedPassword) {
+    @Override
+    public boolean match(String password, String plain) {
         try {
-            final String[] hashedPasswordInfo = hashedPassword.split(HASHED_PASSWORD_DELIMITER);
+            final String[] hashedPasswordInfo = password.split(HASHED_PASSWORD_DELIMITER);
             final SecretKeyFactory factory = SecretKeyFactory.getInstance(hashedPasswordInfo[0]);
             final byte[] salt = DatatypeConverter.parseHexBinary(hashedPasswordInfo[1]);
             final int iterationCount = Integer.parseInt(hashedPasswordInfo[2]);
             final int keyLength = Integer.parseInt(hashedPasswordInfo[3]);
 
-            final KeySpec spec = new PBEKeySpec(passwordInput.toCharArray(), salt, iterationCount, keyLength);
+            final KeySpec spec = new PBEKeySpec(plain.toCharArray(), salt, iterationCount, keyLength);
             final byte[] hash = factory.generateSecret(spec).getEncoded();
             return hashedPasswordInfo[4].equals(toHex(hash));
         } catch (NoSuchAlgorithmException | InvalidKeySpecException e) {
@@ -55,7 +58,7 @@ public class HashUtil {
         }
     }
 
-    private static String toHex(byte[] salt) {
+    private String toHex(byte[] salt) {
         return DatatypeConverter.printHexBinary(salt);
     }
 }
